Fix warning
